# [numpy] add op tri