--- a/python/mxnet/numpy/fallback.py
+++ b/python/mxnet/numpy/fallback.py
@@ -101,7 +101,6 @@ __all__ = [
     'trapz',
     'tril_indices_from',
     'trim_zeros',
-    'triu_indices_from',
     'union1d',
     'unpackbits',
     'unwrap',
@@ -191,7 +190,6 @@ take_along_axis = onp.take_along_axis
 trapz = onp.trapz
 tril_indices_from = onp.tril_indices_from
 trim_zeros = onp.trim_zeros
-triu_indices_from = onp.triu_indices_from
 union1d = onp.union1d
 unpackbits = onp.unpackbits
 unwrap = onp.unwrap